_stbt.core._AnnotatedTemplate.friendly_name: Return relative path
This is the path given by the caller of `stbt.match` (et al), instead of
an absolute path.

This is so that the contents of the stbt-debug html output don't depend
on the path where you ran the tests from. It's more ambiguous but it
allows the stbt-debug characterisation tests to work regardless of the
location of your checkout.

This also affects:

- The video annotations for stbt.wait_for_match.
- The debug output from stbt.detect_match: "Searching for
  {friendly_name}".
- The debug output from stbt.wait_for_match: "Matched {friendly_name}".
- MatchTimeout.expected parameter.
  - Also the str representation of MatchTimeout: "Didn't find match for
    '{expected}' within {timeout_secs}".